requirements.txt
flask>=2.3.0
google-generativeai>=0.3.0
Pillow>=9.5.0
PyPDF2>=3.0.0
werkzeug>=2.3.0
pyngrok>=7.0.0
nest-asyncio>=1.5.0
python-dotenv>=1.0.0
flask-cors>=4.0.0